requirements.txt
pyscreenshot==0.5.1
Pillow==3.1.2
PySSTV==0.4.1
